WIP: Add banhammer button to thread view
Hey Dries,

This is a really poor demo implementation of a button I'd like to have on the main user UI in order for mods to more readily delete spam threads (rather than searching through the admin console.  Any thoughts on if this button would be tolerable here?

Note: If you try this it is probably going to look terrible - it's strictly a mockup.  DO. NOT. MERGE. AS. IS.  :)